Refactored some code to remove unnecessary calls to a redundant composable
