+++ app/src/main/java/com/example/expensetracker/data/category/CategoriesRepository.kt
@@ -4,12 +4,14 @@ import com.example.expensetracker.model.Category
 import kotlinx.coroutines.flow.Flow
 
 interface CategoriesRepository {
+    fun getCategoriesStream(categId: Int): Flow<Category?>
     fun getAllCategoriesStream(): Flow<List<Category>>
     fun getAllActiveCategoriesStream() : Flow<List<Category>>
-    fun getCategoriesStream(categId: Int): Flow<Category?>
-    fun getCategoriesFromTypeStream(categId: Int): Flow<List<Category>>
 
     suspend fun insertCategory(category: Category)
     suspend fun deleteCategory(category: Category)
-    suspend fun updateCategory(account: Category)
+    suspend fun updateCategory(category: Category)
+
+    fun getAllParentCategories() : Flow<List<Category>>
+    fun getAllSubCategories() : Flow<List<Category>>
 }
