+++ app/src/main/java/com/example/expensetracker/model/Category.kt
@@ -17,3 +17,38 @@ data class Category(
     val active : Int = 0,
     val parentId : Int = 0
 )
+
+
+
+//Data class for CategoryUiState
+data class CategoryUiState(
+    val categoryDetails: CategoryDetails = CategoryDetails(active = "true", parentId = "", categName = ""),
+    val isEntryValid: Boolean = false
+)
+//Data class for CategoryDetails
+data class CategoryDetails(
+    val categId : Int = 0,
+    var categName : String,
+    val active : String = "1",
+    val parentId : String = "-1"
+)
+
+// Extension functions to convert between [Category], [CategoryUiState], and [CategoryDetails]
+fun CategoryDetails.toCategory(): Category = Category(
+    categId = categId,
+    categName = categName,
+    active = active.toInt(),
+    parentId = parentId.toInt()
+)
+
+fun Category.toCategoryUiState(isEntryValid: Boolean = false): CategoryUiState = CategoryUiState(
+    categoryDetails = this.toCategoryDetails(),
+    isEntryValid = isEntryValid
+)
+
+fun Category.toCategoryDetails(): CategoryDetails = CategoryDetails(
+    categId = categId,
+    categName = categName,
+    active = active.toString(),
+    parentId = parentId.toString()
+)
