+++ app/src/main/java/com/example/expensetracker/ui/screen/operations/entity/category/CategoryEntryViewModel.kt
@@ -9,61 +9,4 @@ import com.example.expensetracker.data.category.CategoriesRepository
 import com.example.expensetracker.model.Category
 
 class CategoryEntryViewModel(private val categoriesRepository: CategoriesRepository) : ViewModel() {
-    var categoryUiState by mutableStateOf(CategoryUiState())
-        private set
-
-    fun updateUiState(categoryDetails: CategoryDetails) {
-        categoryUiState =
-            CategoryUiState(categoryDetails = categoryDetails, isEntryValid = validateInput(categoryDetails))
-    }
-
-    suspend fun saveCategory() {
-        Log.d("DEBUG", "saveCategory: Called!")
-        if(validateInput()) {
-            Log.d("DEBUG", "saveCategory: Input Valid!")
-            categoriesRepository.insertCategory(categoryUiState.categoryDetails.toCategory())
-        }
-    }
-
-    private fun validateInput(uiState: CategoryDetails = categoryUiState.categoryDetails): Boolean {
-        Log.d("DEBUG", "validateInput: Validation Begins!")
-        Log.d("DEBUG", uiState.categName)
-        return with(uiState) {
-            categName.isNotBlank()
-        }
-    }
 }
-
-//Data class for CategoryUiState
-data class CategoryUiState(
-    val categoryDetails: CategoryDetails = CategoryDetails(active = "true", parentId = "", categName = ""),
-    val isEntryValid: Boolean = false
-)
-//Data class for CategoryDetails
-data class CategoryDetails(
-    val categId : Int = 0,
-    var categName : String,
-    val active : String = "1",
-    val parentId : String = "-1"
-)
-
-
-// Extension functions to convert between [Category], [CategoryUiState], and [CategoryDetails]
-fun CategoryDetails.toCategory(): Category = Category(
-    categId = categId,
-    categName = categName,
-    active = active.toInt(),
-    parentId = parentId.toInt()
-)
-
-fun Category.toCategoryUiState(isEntryValid: Boolean = false): CategoryUiState = CategoryUiState(
-    categoryDetails = this.toCategoryDetails(),
-    isEntryValid = isEntryValid
-)
-
-fun Category.toCategoryDetails(): CategoryDetails = CategoryDetails(
-    categId = categId,
-    categName = categName,
-    active = active.toString(),
-    parentId = parentId.toString()
-)
